refactor: reorganize prediction and database tests
